Add files via upload
This R script generates simulated datasets for time-to-event outcomes in which the population comprises both susceptible and immune to the event of interest.
The susceptible individuals are associated with either the time to the event occurrence or the censoring time (both are random variables), whichever comes first.
The immune individuals are associated with a stochastic time to curation identification or the censoring time (both are random variables), whichever comes first. 

This dataset reflects situations like the time to death during hospitalization, in which immune individuals leave the hospital without dying (curation identification). Clearly, both the time to death during hospitalization and the time until hospital leave are random variables.